Change MPI_Abort calls into halts.
Calling MPI_Abort can cause things printed before it to be lost. So a pattern
like 'writeln("bad setting for foo"); MPI_Abort();' would lose the error
message and just print a partial callstack and a message saying MPI aborted
without giving a good indication of where the problem is.

Replace those calls with halts so that the intended error gets printed and
the line and filename information is printed as well.

Also, update some error messages to suggest setting an environment variable
to fix the problem.