Compute nightly memleaks date once instead of on the fly
The memleaks log file looks like `memleaks.YYYY-MM-DD.nightlyfull.log`.
Previously, we computed the date dynamically each time `memleaks_log()`
was called, but this could cause problems when testing ran past a new
day since at the end we'd try to copy a file for the current day instead
of the day testing started.

To fix this just compute the date once when testing starts.