Greatly simplify the check on if a use is visible during function resolution
We were performing an extensive check to see if a particular use was visible
from the call site.  However, for the most part, if a use was private the answer
would be that it was not visible, and it was very likely to be not visible if
we had reached that use through another use.

This change makes it so that we just ignore private uses if we reached the use
statement through a use chain.  If we reached the use through going up scopes,
we don't care if it is public or private since we are in the scope that defined
it.

Note that there is an edge case here that is important: when a use chain would
bring us back to a scope that also happens to be a parent scope of the call and
that scope has a private use.  We'll reach it first from the use chain and so
mark the scope as seen, even though we won't have traversed the private uses.
I plan to see what fails as a result and if nothing does, add a test of the
case.  If something fails, I'll just add the additional check.