Add second traversal in case of parent scopes being accessed via uses
It could happen that a used module has a use of a parent scope, causing it to
be found in such a way that we ignore its private uses.  When traversing up
scopes, if we find that we have already seen a module, look through only its
private uses, to ensure that we don't miss anything.

TODO: I think I can simplify this further, but I want to make sure it resolves
the correctness issues first.