Don't test comd-elegant with comm=ofi.
Similar to other configurations, with comm=ofi comd-elegant's spawn()
call causes trouble, at least with the 'verbs;ofi-rxm' provider in ml-16
perf testing.  Skip this test for comm=ofi for now, until we can better
characterize just what the problem is and either figure out a way to
deal with it or refine the skipping.