Support the bytes type in multilocale interop (#14471)
Support the bytes type in multilocale interop (#14471)

This PR adds support for the bytes type to multilocale libraries. It
also refactors some code in `compiler/codegen/mli.cpp` to replace
uses of `and` and `or` with `&&` and `||` (I was unaware of style
conventions for the compiler when I first wrote that code).

Currently, the Chapel server in a multilocale library will leak the
memory allocated for aggregate (string, bytes) return values. This
is a bug that will hopefully be fixed in the near future. I tried to
introduce some code to deallocate return values but was met with
segfaults at runtime for some tests.

I've opened a private issue about the leak for now.

---

Testing:

- [x] ALL on linux64 when CHPL_COMM=none
- [x] ALL on linux64 when CHPL_COMM=gasnet (no cygwin/numpy)
- [x] test/interop/python on linux64 when CHPL_COMM=gasnet

---

Thanks to @lydia-duncan for review!