Add multi-locale perf testing for NPB FT
This tests performance of NPB-FT with and without `usePrimitiveComm`

This also adds `-I$FFTW_INC -L$FFTW_DIR` to the compopts in order to get
this test working on systems that don't use the craype compiler
wrappers.

Lastly this fixes single-locale correctness testing by filtering out the
number of locales in the prediff.