Make reviewer suggested changes to HDF5 single-file writer
Updated doc comments, changed asserts into conditional halts, added a check
and compiler error that the array is block distributed. Move some extern
declarations into the function that uses them.

Update a test to both write and read a HDF5 file and verify the result is
what is expected. Add a file it generates to the .gitignore file.

Also added some more close calls that were missing at the end of the function.