Change test to use C99 portable format specifier (#14501)
Change test to use C99 portable format specifier (#14501)

This PR adjusts a test to use a portable C99 format specifier in a
printf call. In #14471 I changed the bytes wrapper type to use
fixed width scalars for portability, and forgot to adjust this test.

---

Testing:

- [x] interop/C/exportBytes/noArgsRetBytes.test.c on darwin
- [x] interop/C/exportBytes/noArgsRetBytes.test.c on linux64

---

Change to test, not reviewed.