Get comm=ofi perf testing working.
Adjust nightly testing support scripts to add regular comm=ofi perf
testing on Cray CS, via test-perf.cray-cs.ofi.bash which will be invoked
through Jenkins.  Related to this, factor out script code common between
this and the comm=gasnet perf testing that already runs on Cray CS, and
put that common code in common-cray-cs.bash or common-perf-cray-cs.bash,
as appropriate.  While here, also refactor common-ofi.bash so that it
can be used for both comm=ofi functional and performance testing.

Also while here, change most uses of the `[ ]` bash builtin in these
scripts to `[[ ]]`, because I was using the former but expecting the
`==` and `!=` operators in it to do filename pattern matching, and
that actually just does string matching.  This resulted in some spurious
settings, now corrected.