Suppress Fortran unhandledType test until LLVM munges internal symbols
This test is sensitive to the names of symbols in the generated code
and required updating when I started munging internal symbols in PR
symbol munging for LLVM last night, causing it to fail in LLVM test
configurations.  Here, I'm suppressing the failure when --llvm is
thrown to address this and to help remember to re-enable it when LLVM
isnt' special-cased in this way.