Eliminate spurious "this source location is a guess" in warnings
In some cases the compiler produces this output for a compilerWarning() call:

    LOCATION: warning: TEXT
    Note: This source location is a guess.
    LOCATION: warning: TEXT

where LOCATION and TEXT are determined by the call.
This PR changes it to just:

    LOCATION: warning: TEXT

It updates .good files correspondingly.

BACKGROUND

The compiler uses caching when issuing warnings for compilerWarning() calls
in Chapel code. There is a good reason for it. The compiler prints a
warning - by invoking USR_WARN - when it encounters __primitive("warning")
while resolving a compilerWarning function. This happens only the first time
that compilerWarning() is invoked with a given set of actual arguments.
If there is a compilerWarning() call with the same set of actuals in a
different point in the code, compilerWarning() is not resolved again.
Instead, the compiler checks the cache and invokes USR_WARN when there is
a cache entry.

What if there are several different paths to the compilerWarning() call that
share a suffix? ex. `a() -> x() -> y() -> compilerWarning("example")` and
`b() -> x() -> y() -> compilerWarning("example")` ? To issue the warning,
the cache needs entries for the appropriate instantiations of compilerWarning(),
y(), and x().

The current implementation of caching, based on `innerMap` and `outerMap`,
is not perfect because it works incorrectly in some corner cases and because
it can generate the same warning twice, doubting itself the first time by
printing "Note: This source location is a guess."

The current PR does not address the imperfection. It simply improves user
experience by recognizing the latter scenario happens and fixing it up.