User error, not crash, for compilerWarning(string,bool)
This converts a call like `compilerWarning("hi", true)` to a user error
"bool arguments must be cast to string".

BACKGROUND

When a function has a param int formal, a param boolean can be passed to it,
as expected, because of to implicit conversion bool->int.
While resolving that function, however, while the formal's type is `int`,
`paramMap` has the boolean value for that formal. This used to crash
`resolveExprMaybeIssueError()` in the compiler. This mismatch might get us
in trouble in other corner cases too, although I am unaware of that.