Check the type argument of numFields() et al.
Make it a user error, not a compiler crash, when a non-record/class/union
type is passed to Reflection.numFields() and the other field-query functions
in the Reflection module. The types that are implemented internally
using records however are not Chapel records result in such an error as well.

Also, for a class type, unwrap the _owned/_shared wrapper if applicable.
This augments existing unwrapping in preFold.cpp for PRIM_NUM_FIELDS et al.

Given that DefaultAssociative.chpl uses numFields() and getField() on ranges,
I created internal counterparts of these that accept ranges.
numImplementationFields() and getImplementationField() still need to be
protected from an arbitrary type argument because otherwise they would
result in a compiler crash. The set of permitted argument types for these
can be extended by editing isImplementedWithRecords().