Access param, type fields without postfix-!
With this change, the postfix-`!` operator is no longer needed
when accessing param and type fields of values of nilable class types.

* This does not include user-defined accessor methods.
Postfix-`!` is still required as before when invoking these.

* This does not include type fields instantiated with runtime types.
Postfix-`!` is still required as before to access these.
Since at runtime such fields still carry values, they behave
like var/const fields w.r.t. nilability.

* While there, add a compiler error when accessing a type field
of a class type when the field is instantiated with runtime types.
Otherwise the generated code dereferences a null pointer.
This is because class types are not considered runtime types in the
compiler, so do not carry runtime type information. See #11549.

* This change also implements accesses to param/type fields of `owned`
and `shared` class types and values.

While there:
* viewFlags() to print "type method" when appropriate, which is useful.
* wrapAndCleanUpActuals() to take actualIdxToFormal by ref, for consistency.
* Add a future test for an existing issue.

The implementation is based on making the receiver of the compiler-generated
param/type field accessor generic over the management strategy and nilability.
This was already the case for accessors to types, here I extend it to
accessors to values. This happens during buildDefaultFunctions, so the accessor
needs to include hooks that will fire during resolution.

I considered also an alternative implementation where I made
allowImplicitNilabilityRemoval() return true when resolving such a field
accessor. The show-stopper for that was that the compiler would insert a cast
to convert a nilable actual to the non-nilable borrow receiver formal
of the compiler-generated field accessor. Such a cast "throws", resulting
in a "throw without catch" error during lowerErrorHandling().
The advantage of this alternative is that it happens during resolution,
so the type of the receiver and the type of the field are known.
For the record, if we handled these casts somehow, we would need also
the following adjustments. (A) Flag compiler-generated accessors with
FLAG_COMPILER_GENERATED. This is so that user-defined accessors can be
excluded from the special treatment. (B) Exclude FLAG_FIELD_ACCESSOR-flagged
functions from `compilerGenerated` in lateConstCheck.cpp / printReason().
Otherwise field access will be omitted from the list of reasons
under --no-devel in these tests:
functions/ferguson/ref-pair/record-this-intent-const-error
parallel/forall/vass/other/fields-of-this-2-const-var
parallel/taskPar/taskIntents/fields-of-this-2-const-var
types/records/const-checking/this-in-method