Deprecate postfix-! on types
Now the compiler issues a deprecation warning when the postfix-`!` operator
is invoked on a type. The warning also suggests using instead a cast to
`class` or to `borrowed class`. I offer both alternatives because I think
they are both valuable in different ways: a cast to `class` simply removes
nilability; a cast to `borrowed class` is closest to the current semantics
of postfix-`!` on types.

The tests are adjusted to cast to `class` where that is sufficient,
otherwise to `borrowed class`. A couple of exceptions:

* Move this test into `test/deprecated` :

    test/classes/nilability/type-bang.chpl

* Leave these tests as-is - updating their .good - because they test
specifically the bang operator in the context of a cast. To revisit
once we remove the for-now-deprecated feature.

    test/parsing/precedence/bang-vs-cast.chpl
    test/parsing/precedence/bang-vs-new-etc.chpl