Extends UseStmt to store the used module's new name and some helper methods
Adds a new field modRename to the UseStmt and sets it via the constructor.  All
calls to the constructor have currently been updated so that they just send ""
to the call, but when the parser has been modified real values will be sent.

Also adds two methods to the type: an overload of isARename that takes no
argument (indicating that we are looking for the rename of the module being
used), and an overload of getRename that takes no argument (and thus returns the
rename of the module being used).  These methods are not currently used, but I
intend to extend scope resolution and function resolution to call them.