Preserve the line number when re-throwing an exception (#14570)
Preserve the line number when re-throwing an exception (#14570)

This PR adjusts error handling so that thrown Errors only have their
file and line info set the first time they are thrown, rather than
every time they are thrown.

It adds a bool flag to the Error class and uses that internally so
that calls to `chpl_fix_thrown_error` inserted by the compiler
only set the file/line values of the Error once.

Resolves #11300.

---

I chose this approach because it is simple. A side effect of this
approach is that the compiler is still inserting redundant calls
to `chpl_fix_thrown_error` that essentially do nothing, and that
Errors take up a bit more space.

In an ideal world we would either tie getting file/line info to an
Error's initializer, -OR- have the compiler remove the redundant
calls to `chpl_fix_thrown_error`.

This is a low investment solution in the meantime.

---

Testing:

- [x] ALL on linux64 when CHPL_COMM=none
- [x] ALL on linux64 when CHPL_COMM=gasnet

---

Thanks to @vasslitvinov for review!