Remove deprecated "array-as-vector" routines on arrays (#14557)
Remove deprecated "array-as-vector" routines on arrays (#14557)

This PR removes deprecated "array-as-vec" routines from arrays along
with their tests in `test/deprecated`. It also adds three new tests,
two of which are futures related to lifetime checks that worked for
`array.push_back` but do not work for `list.append` (see #13613).

Resolves #14533.

---

Testing:

- [x] ALL on linux64 when CHPL_COMM=none, CHPL_LLVM=none
- [x] ALL on linux64 when CHPL_COMM=gasnet, CHPL_LLVM=none

---

Thanks to @daviditen for review!