Improve the message comm=ofi prints for FI_EMFILE.
We get the libfabric -FI_EMFILE error code when it needs to open a file
but we've reached the open file limit.  Basically it's the same thing as
the EMFILE errno code.  Here, improve our handling of this one: don't
only print it for the 'tcp' provider, and also don't only print it when
some computation based on numLocales and comm concurrency is larger than
the open file limit.  We're just not smart enough to get that right in
all cases.