Have extern blocks inject `use SysCTypes;` at same scope
This fixes a regression introduced in PR #14524 in which, now that
SysCTypes is no longer auto-used, extern blocks break since they
relied on the C types being automatically available.

The approach taken here is to have the parser's build routine that
generates a new extern block also insert a `[public] use SysCTypes;`
statement just before it using the same pattern that the parser uses
to build that statement if it were in user code.