Improve testing error message when compilation fails
When compilation fails and there's no <basename>.good or compopt
specific .good file to match against, issue a "compilation failed" error
message instead of saying "there was no .good file". This should make
it more obvious that there was a hard failure rather than just a lack of
a testing file. This is an extension of 4376.