Update package modules w.r.t. public uses of HDFS
This PR updates the set of package modules whose interfaces require
C types as defined in SysCTypes to include a `public use SysCTypes`
so that clients of the modules will not have to use SysCTypes in
order to make calls into them.  Ideally, it would be nice if any
Chapel-facing code in these modules did not require C types and
that any conversion to C took place under the hood if need be, but
that was a bigger change than I was able to take on here.  And
arguably, since they're just package modules, it's not so big of a
deal.  But we might consider making this an expectation for a given
module before it transition to the standard modules directory.