Add tests ensuring that package modules can compile
A common cause of surprises in nightly testing is that
package modules that aren't tested in vanilla linux/Mac
configurations blow up only in the configurations that
test them—i.e., those for which a given library is
available (or assumed to be available).

This PR attempts to address this by creating tests that
ensure that we can codegen programs that use the packages
without trying to link them.  In this way if anything
goes wrong at the Chapel level, we'll know about it in
one of those vanilla test runs.

The approach I took here was to

* symbolically link to a test that exercises the package when we test
  that package
* have the compiler stop compiling after the codegen pass via COMPOPTS
* skip execution via NOEXEC
* make sure the output is quiet via empty .good files

Note that this is not bulletproof by any means.  If the package
in question runs into problems only when some other routines in
it are called, then these tests won't catch it.  However, this
should head off some general catastrophic failures that could
be avoided in any use of the module which are the kind I've been
running into most often while trying to lock down the default
namespace better.