Allow messages in new Errors (#14545)
Allow messages in new Errors (#14545)

This is a simple PR that enables a user to attach a message string to
the Error base class. It also adds one new test to lock in the new
behavior.

Users can call an initializer overload that takes a `string` argument
in order to attach a message that is displayed when an error is
printed or when an unhanded exception occurs.

This PR is motivated by extensive discussion in #9584.

Resolves #9584.

---

Testing:

- [x] ALL on linux64 when CHPL_COMM=none, CHPL_LLVM=none
- [x] ALL on linux64 when CHPL_COMM=gasnet, CHPL_LLVM=none

---

Thanks to @ben-albrecht for review!