Add standalone version of dynamic() iterator
While talking with Tiago about parallel iterators in Chapel, I'd hoped
to show him a standalone version of the dynamic iterator as a familiar
example, but it turned out that we didn't have one.  This PR adds one
for the range version of dynamic().  In the future, we should consider
adding them for the domain version as well as for the other iterators
in this module.