Fix a 32-bit sensitivity with the I/O plug-in
The addition of my compile-but-don't-link package tests revealed a
portability issue on 32-bit platforms in which we were relying on
ssize_t and int(64) being the same type.  Here, I'm changing a
variable of type ssize_t to int(64) because it seems that all calls to
which its passed are expecting an int(64).