Make the lastexecopts really be last.
When we use lastexecopts, we really need those to be the last thing on
the command line sub_test runs.  Typically this is because the test is
checking that a partial command line option, such as '-sconfigVar='
without a value to assign, generates the right error message.  But when
system launcher time limiting is in use, as with launchers containing
any of "pbs", "qsub", or "slurm in their names, we've been adding the
'--walltime' option after the lastexecopts.  Doing so breaks tests with
.lastexecopts files.  Here, move the system launcher walltime limiting
option so that it precedes the lastexecopts.