Don't run atexit(3) functions if we think that might cause a segfault.
Paraphrasing the comment added by this commit: if exiting unilaterally
with the 'verbs' provider in use, call _exit() instead of allowing the
usual runtime control flow to call exit() and invoke the atexit(3)
functions.  Otherwise we run the risk of segfaulting due to some broken
destructor code in librdmacm.  That was fixed years ago but the fix
seems not to have made it into general circulation yet.