Remove superfluous and poorly supported exit(err).
This test does exit(err) at the end of execution on the last node,
apparently to make its exit status be the number of errors seen.  But
this is superfluous in terms of its pass/fail state, because the .good
file matches iff the number of errors is 0 (zero) anyway.

In addition, the exit(err) is problematic because exit() in a Chapel
program may have system-specific side effects.  As examples, on one of
our Cray XC systems with the slurm-srun launcher and comm=ugni this
program hangs after printing output that would match the .good file
because the rest of the nodes just sit there waiting for something to
happen, while on one of our CS systems with the slurm-srun launcher and
comm=ofi the PMIx client launcher support library apparently kills the
rest of the nodes when it sees the one that executes exit(err) do so,
and then slurm produces some extra output saying that it saw that
happen.  These individual cases could be handled, of course, but others
would no doubt arise.

So, since the exit(err) is both unnecessary for correctness checking and
hard to deal with, here we get rid of it.