dynamic() iter: coordinate via chunk index instead of chunk offset
This avoids overflowing both the calculation of high and the state
of curIndex when the end of the range is within chunkSize of the
max of the index type.

There's no reason to adjust high down to remain.high, since we're
about to apply range intersection with remain.  This lets high be
const.

low will always be <= high at the end of the loop body so no need
to test it any more.