Generalize and simplify the task callback tests.
The main change here is to adjust the callback reporting to remove 'end'
entries for tasks whose 'create' and 'begin' entries we didn't see.  In
the particular case I'm seeing with comm=ofi, during the time we have
the callbacks registered we see the 'end' for a task-wrapped GET Active
Message (which must have been to access unregistered memory), where the
task involved must have started before we registered for callbacks.

While here I did a couple of other minor cleanups.