Extend nilChecking to account for if-statements
Resolves #11918. Removes some unnecessary calls to postfix-`!`.
Also addresses #13630, I plan to close it with a separate PR.

The nilChecking pass now understands the effect of conditionals that compare
a class variable against `nil`.

* Within the then/else branches of such a conditional, the variable is known
to start out as nil or non-nil, as appropriate.

* If the branch will never be executed at runtime, nilChecking does not
analyze it and so does not issue "you are dereferencing nil" errors
that it would otherwise.

* Calls to postfix-`!` that are known to succeed or never executed
are removed. This way, they do not incur runtime cost and do not
potentially disable some optimizations.

For example:

    var x: owned MyClass?;
    var y = ....;

    if x != nil then
      writeln(y!);     // never executed, an error is never issued

    if y != nil then
      writeln(y!);     // the runtime ! check is removed

This does not address field accesses, partly because it is hard for
the compiler to ensure that the field is not modified concurrently.

Next steps: analyze conditionals like these:

    var x = ...;
    if x == nil then
      x = new MyClass();  // (a)
      return;             // (b)
      exit();             // (c)

    writeln(x!);          // x is never nil if (a), (b), or (c) happens

Once we eliminate holes in our compile-time checking of nilable types,
we can also remove `!` on formals of non-nilable types.

Minor: gatherVariablesToCheck() right now serves only to detect the
functions that have no class variables to be analyzed. We can simplify it
to do only that.

Implementation Notes

I considered these approaches to handling unreachable branches:

* Remove them from the AST altogether.
* Do not analyze; remove postfix-! calls from them.
* Analyze normally, as if they are reachable.

I chose the second option as a middle ground. The first option is aggressive
and I did not have enough confidence in the correctness of nilChecking.
The situation when a branch is known to be unreachable due to a nil check
is a corner case, which is hard to test thoroughly.

The third option is the most conservative. I chose against it because
it comes with this challenge:

    var x = nil;

    if x != nil then
      // in the then-branch the compiler knows x!=nil
    else
      // here the compiler knows x == nil

    // To the compiler, since x!=nil in one branch and x==nil in the other branch,
    // after the conditional x can be either after. So x==nil is no longer known.

The nilChecking code I added propagates the "is non/nil" information along the
use-def chains through casts. This is because, prior to testing `if x != nil`,
the compiler casts `x` to `object` if it is not already.

Minor: I slightly tuned up the debugging output.

Testing changes: I combined the two versions of `if-object-2` into one
and moved it next to `if-object-1.chpl`.