Convert Chapel strings to Python strings, support embedded nulls (#14592)
Convert Chapel strings to Python strings, support embedded
nulls (#14592)

This PR adjusts Python interop so that Chapel strings returned from
exported routines are converted to Python strings instead of
Python `bytes`.

Previously, Chapel strings were returned to Python as `bytes`, and
only Python `bytes` could be passed to exported Chapel routines that
had string parameters.

The Chapel string now maps cleanly to the Python string in Python
interop. Additionally, this PR also supports passing strings that
contain null bytes between Chapel and Python.

Previously, we used `strlen` to get the length of a Chapel string
(as it was exported to C as a `char*`), but now both strings and
bytes are shuffled around between languages as a `chpl_byte_buffer`.

As the name implies, `chpl_byte_buffer` is a buffer of bytes with an
explicit size, enabling strings with embedded nulls to be passed
comfortably between languages.

---

Testing:

- [x] ALL on linux64 when CHPL_COMM=none (no cython/numpy)
- [x] interop on linux64 when CHPL_COMM=none
- [x] interop on linux64 when CHPL_COMM=gasnet
- [x] interop on darwin when CHPL_COMM=none

---

Thanks to @lydia-duncan for review!