Add a basic PALS-based launcher.
This adds a basic Chapel launcher for the Cray PALS system launcher on
Shasta systems.

In support of this, I also added an argv-builder function to the common
launcher code, adapted from existing code in the mpirun4ofi launcher.
This helps avoid the problems with the statically sized buffers and arg
lists which are used in the other launchers.