Migrate array future so that it is only tested under Valgrind (#14675)
Migrate array future so that it is only tested under Valgrind (#14675)

This PR modifies an array future so that it is only tested under
Valgrind. Before this PR, the test `test/arrays/bugs/arrOfTupleOfArr`
expected a nil dereference error as its failure mode.

However, under gasnet configurations, the test seems to fail with a
segfault instead. In #13149 @vasslitvinov suggested that Valgrind
might produce a more precise failure mode for this test.

I've modified this test to be skipped if `CHPL_TEST_VGRND_EXE != on`,
and added a prediff / changed the `.bad` file accordingly.

Thanks to @e-kayrakli for review!