Remove --legacy-classes from two .compopts
This furthers the goal of eliminating `--legacy-classes` compiler flag.

This PR modifies these two tests:

    test/classes/diten/callNilClassesMethod.chpl
    test/classes/diten/nilDynamicDispatch.chpl

They used to require --legacy-classes to have a variable of a non-nilable
class type that holds `nil`. Now they use a hole in typechecking of
nilable types to achieve that. Specifically, right now array elements
can be `nil` despite having a non-nilable type.

When this hole is fixed, we can switch to using another hole.
When no holes remain, we can remove altogether the nil-checking being tested
here, and therefore these tests as well.

BACKGROUND

These tests' purpose is to test "old-style" runtime nil checking, which applies
to the receiver upon each method call. To observe it fire, we need
a class variable holding `nil`.

Alas, if this variable has a nilable type, the old-style check will not fire
because:
* the typechecker will insist on calling postfix-`!` on the variable
  before it is passed to a method, and
* postfix-`!` will fire its own error, so the old-style check is never
  executed.

Therefore we need the variable to have a non-nilable type, so that
the typechecker does not insist on calling postfix-`!`.
This change relies on a loophole in the typechecker to make it happen.
Which loophole is used does not matter.