Add dynamic-int-overflow and dynamic-uint8-overflow tests
dynamic-int-overflow tests for the problem in #14618.  I set a 7
second timeout because even when the problem is present, the test
would evetually complete with no indication of the problem besides the
runtime.  We can't print anything (or do any work that --fast won't
optimize out) because there are too many loop iterations to complete
when there's no bug.

dynamic-uint8-overflow is a test for the same problem using uint(8).
The artimetic's easier to follow, and it runs completely without
--fast whether or not the bug is present.  But it doesn't run on
master prior to this PR because of #14617.