Non-nilable variables to be non-nil in nilChecking
This PR is a follow-up to #14660. It modifies nilChecking so it treats variables
of non-nilable types as "known to contain non-nil" -- in some cases.
Specifically:
* the formals of non-nilable types
* results of calls to postfix-! (singled out because it is easy to check)
* results of calls to functions whose return types are non-nilable

Due to loopholes currently in typechecking of non-nilable types,
this is not sound. So nilChecking is also modified so it no longer removes
ALL postfix-! calls in a conditional statement's branch that is known to the
analysis as unreachable. nilChecking continues removing those postfix-! calls
where the argument is known as non-nil. If the argument is actually nil at
runtime, due to a current typechecking loophole, the old-style runtime nil check
- still inserted by the compiler regardless of the outcomes of nilChecking -
will fire and halt the program, instead of postfix-! halting the program.
Note that regardless of this change, postfix-! does not include the runtime
check when the argument type is non-nilable.

IMPLEMENTATION DETAILS

Detection of non-/nilability of a type during the nilChecking pass cannot be
done the way it is done during resolution, i.e. using isNonNilableClassType().
This is because for `owned` and `shared` types the latter involved checking
the type of the owned/shared record's field `chpl_t`. Since this is a type
field, it is removed at the end of resolution. To communicate non-nilability
of owned/shared types to nilChecking, this PR introduces
FLAG_MANAGED_POINTER_NONNILABLE.

Suggested by/discussed with @mppf .