Fix problem with forall primer
Forall loops have several DefExprs that are within the ForallStmt.
At first I had trouble identifying these because they are not printed
out in the AST log (so I fixed that). Then, these variables are not
visited by walkBlock/walkForallBlocks in addAutoDestroyCalls and as
a result never have deinit calls added for them. We might consider
adding a FLAG_NO_DEINIT to these variables but for the time being
I just turned off the early deinit behavior for them.