Save the point of renaming and the new name on a SymExpr
Prior to this, when a renamed symbol was found through resolution, we'd lose
that it was originally referred to by another name in context, so when the
symbol was used in an error message related to another symbol (e.g. when a
module was renamed and then we failed to correctly access a symbol in that
module through explicit naming), we'd refer to it using its actual name instead
of the name used in context.

This could be confusing to the user, so I updated the code to save this renaming
information.  Then, during `resolveModuleCall`, I updated the code to use the
new name for the module in the error messages and added a note with the actual
module name in it (so it is easier for the user to find the source module)