Remove a redundant and out-of-place copyright in the online spec
While doing other spec edits this week, I noticed that this copyright
was a bit of a stumbling block and redundant with the ones at the
bottom of the page, so am removing it.  There's arguably more we
should do to make the front page of the online spec more descriptive
and easier to approach, but this was low-hanging fruit in the
meantime.