Make the <~> operator equivalent to `channel.read`/`channel.write` (#14430)
Make the <~> operator equivalent to `channel.read`/`channel.write`
(#14430)

This PR makes the <~> IO operator throw, and insofar as possible
equivalent in semantics to `channel.read` and `channel.write`.

There are some unavoidable differences between the two - i.e., the
`channel.read` routine opts to return `false` if it hits the end
of a file, instead of throwing a SystemError.

If a SystemError is thrown from within IO routines up through
`readThis`/`writeThis`, then the error is caught and the equivalent
QIO error code is stored in the channel. If the user throws an
error that is not a SystemError from within their overload of
`readThis`/`writeThis`, the EIO error code is stored in the channel
as a stopgap to indicate an error has occurred.

As of this time there is no way to distinguish between SystemError
thrown by the user and SystemError that are thrown by the Chapel
IO machinery.

This PR adds the `throws` keyword to all the `dsi` IO routines for
arrays, as a consequence of modifying <~>.

The test `test/classes/deitz/class/tree.chpl` was rewritten to use
an iterative postorder traversal instead of a recursive iterator
due to compilation bugs. See #7134 for details.

---

Testing:

- [x] ALL on linux64 when CHPL_COMM=none
- [x] ALL on linux64 when CHPL_COMM=gasnet

---

Thanks to @mppf for review!