Update copyrights in test directory w.r.t. Cray->HPE
It was tough to decide what to do with these Copyrights w.r.t. the
Cray->HPE change because most code in the test directory doesn't carry
a copyright, on the arguments that (a) it clutters up things that are
meant to be user-facing / instructional and (b) most of the code in
the test directory doesn't represent major IP (and even if when does,
arguably it's covered by the repo-wide LICENSE/COPYRIGHT files).

Here, I considered each Cray copyright in the test directory and
took the following actions:

* Updated the copyright to include 2020 and HPE for the following
  files because they seemed to represent a reasonable investment of
  IP:

  - test/library/packages/Itertools/Itertools.chpl
  - test/library/packages/csv/CSV.chpl
  - test/library/standard/DataFrames/DataFrames.chpl
  - test/users/Deeksha96/StringAlgorithms.chpl

* Updated the copyright to include 2020 and HPE for the following
  files because I believe they started as code clones of files that
  already carried the copyrights:

  - test/studies/comd/elegant/arrayOfStructs/util/AccumStencilDist.chpl
  - test/users/aroonsharma/MyBlockCyclic.chpl

* Updated the copyright to include 2020 and HPE for the following
  files because they've carried the copyright for quite some time
  and were on the borderline for me:

  - test/library/packages/HDFS/HDFSiterator.chpl

* Removed the copyright for the following files which are fairly
  straightforward and carried them because they started as code
  clones or moves from files that already carried them:

  - test/modules/paths/int/ChapelNumLocales.chpl
  - test/runtime/ferguson/testatomics.test.c